refactor: Event wrapper class, added hooks prototype
